portcullis: move certificate.rs tests to certificate_tests.rs — the module crossed the 2500-line sweep ceiling once the surface, compartment and parity tests all landed on it
